requirements.txt
Django==4.1.5
django-cors-headers==3.14.0
django-countries==7.5.1
django-smart-selects==1.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0

